requirements.txt
streamlit>=1.28.0,<2.0.0
Pillow>=10.0.0,<11.0.0
